pyproject.toml
[project]
name = "gdrive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.3",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
